refactor: extract vaadin-upload logic into reusable mixin
